lib: move s3util out of internal
